+++ R/att_gt.R
@@ -91,6 +91,32 @@
 #' @param anticipation The number of time periods before participating
 #'  in the treatment where units can anticipate participating in the
 #'  treatment and therefore it can affect their untreated potential outcomes
+#' @param base_period Whether to use a "varying" base period or a
+#'  "universal" base period.  Either choice results in the same
+#'  post-treatment estimates of ATT(g,t)'s.  In pre-treatment
+#'  periods, using a varying base period amounts to computing a
+#'  pseudo-ATT in each treatment period by comparing the change
+#'  in outcomes for a particular group relative to its comparison
+#'  group in the pre-treatment periods (i.e., in pre-treatment
+#'  periods this setting computes changes from period t-1 to period
+#'  t, but repeatedly changes the value of t)
+#'
+#'  A universal base period fixes the base period to always be
+#'  (g-delta-1) where delta is the number of periods of
+#'  treatment effect anticipation.  This does not compute
+#'  pseudo-ATT(g,t)'s in pre-treatment periods, but rather
+#'  reports average changes in outcomes from period t to
+#'  (g-\delta-1) for a particular group relative to its comparison
+#'  group.  This is analogous to what is often reported in event
+#'  study regressions.  
+#'
+#'  Using a varying base period results in an estimate of
+#'  ATT(g,t) being reported in the period immediately before
+#'  treatment.  Using a universal base period normalizes the
+#'  estimate in the period right before treatment (or earlier when
+#'  the user allows for anticipation) to be equal to 0, but one
+#'  extra estimate in an earlier period.
+#'  
 #' @references Callaway, Brantly and Sant'Anna, Pedro H. C. "Difference-in-Differences with Multiple Time Periods" Forthcoming at the Journal of Econometrics <https://arxiv.org/abs/1803.09015> (2020).
 #'
 #' @return an [`MP`] object containing all the results for group-time average
@@ -157,6 +183,7 @@ att_gt <- function(yname,
                    biters=1000,
                    clustervars=NULL,
                    est_method="dr",
+                   base_period="varying",
                    print_details=FALSE,
                    pl=FALSE,
                    cores=1) {
@@ -179,6 +206,7 @@ att_gt <- function(yname,
                         biters=biters,
                         clustervars=clustervars,
                         est_method=est_method,
+                        base_period=base_period,
                         print_details=print_details,
                         pl=pl,
                         cores=cores,